Remove space

+++ bit_manipulation/0-binary_to_uint.c
@@ -3,14 +3,11 @@
 
 /**
  * _pow_recursion - returns the value of x raised to the power of y
- *
  * @x: an integer
  * @y: an integer
- *
  * Return: a value
  */
-
-int _pow_recursion(int x, int y)
+unsigned long int _pow_recursion(int x, int y)
 {
 	if (y < 0)
 		return (-1);
@@ -18,6 +15,7 @@ int _pow_recursion(int x, int y)
 		return (1);
 	return (x * _pow_recursion(x, (y - 1)));
 }
+
 /**
  * binary_to_uint - converts a binary number to an unsigned int
  * @b: the binary number
